ci: integrate automated release branch rebase and merge (#11)
This commit integrates the automated rebase and merge workflow for the release branch into the main branch.

The workflow includes:
    - Auto trigger upon successful merge of a pull request into the release branch.
    - An automatic rebase & merge of release onto main gets performed if no merge conflicts are present.
    - GPG signing of commits during the rebase and merge process.
    - Slack notifications to provide updates on the operation status.

A composite action has been created to encapsulate the rebase and merge logic, ensuring modularity and reusability.

A manual script "local/rebase-merge.sh" is provided for handling merge conflicts and performing interactive rebases when necessary.

This integration streamlines the release process and enhances its security and transparency while providing a convenient fallback mechanism for manual intervention.